Create LaunchEC2InstanceswithLaunchTemplates.ps1
This script automates launching 2 EC2 instances using AWS Launch Templates and using instance values to tag the newly launched instances.

This script will automatically launch 2 EC2 instances in your AWS account. The instances are launched using LaunchTemplates that are
defined already and the script is using the launch template names.

The script will also gather the 1st instances private IPv4 address and use that value to create a tag for the 2nd ec2 instance.

Finally, an output of the executed operations will be displayed on the screen to give the executor an overview of operations.